no it is not
git-svn-id: http://yagf.googlecode.com/svn/trunk@128 bb55e26e-657b-11de-982f-55b50260c994

+++ src/qgraphicsinput.cpp
@@ -506,7 +506,7 @@ QPixmap QGraphicsInput::getImage()
     return hasImage ? m_realImage->pixmap() : 0;
 }
 
-const int stdwidth = 2550.;
+const float stdwidth = 2550.;
 
 QPixmap QGraphicsInput::getAdaptedImage()
 {
